# Conflicts:
#	fbi_analysis_etl/src/com/fr/bi/etl/analysis/conf/AnalysisETLPackageSet.java
#	fbi_analysis_etl/src/com/fr/bi/etl/analysis/data/AnalysisETLOperatorFactory.java
#	fbi_conf/src/com/fr/bi/conf/data/source/operator/create/TableMergeOperator.java

+++ fbi_base/src/com/fr/bi/stable/file/ColumnFile.java
@@ -20,7 +20,7 @@ public interface ColumnFile<T> extends IndexFile {
 
     void releaseDetailDataWriter();
 
-    void addDataValue(long row, T value);
+    void addDataValue(int row, T value);
 
 
     CubeGenerator createGroupIndexCreator(BIKey key,
